COOK-2611 Adding queues parameter to celery LWRP
Bind celery listeners to particular queues by providing an array
of queue names to the 'queues' parameter. Defaults to ['celery'].

This change modifies the provider to use the -Q flag with the
list of queues from this parameter at every invocation. Celery
documentation states that 'celery' is the default queue name, so
this ought not to modify behavior for those currently using this LWRP.